requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pandas==2.2.0
numpy==1.26.2
django-bootstrap-datepicker-plus==5.0.4
django-dynamic-formsets==0.0.8
django-widget-tweaks==1.5.0
async-timeout==4.0.1
